# back-end/docs/swagger.yaml
basePath: /api
definitions:
  handlers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  handlers.LoginResponse:
    properties:
      partnerId:
        type: string
      role:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
...
  models.Partner:
    properties:
      _key:
        type: string
      additionalComments:
        type: string
      aiExperience:
        type: string
      annualRevenue:
        type: string
      arangoEngagements:
        type: string
      companyName:
        type: string
      contacts:
        items:
          $ref: '#/definitions/models.PartnerContact'
        type: array
      country:
        type: string
      createdAt:
        type: string
      graphDatabaseFit:
        type: string
      industryVerticalFocus:
        items:
          type: string
        type: array
      keyPartners:
        type: string
      marketFocus:
        items:
          type: string
        type: array
      numberOfEmployees:
        type: string
      primaryPartnerBusiness:
        items:
          type: string
        type: array
      registrationStatus:
        type: string
      statusHistory:
        items:
          $ref: '#/definitions/models.StatusHistoryEntry'
        type: array
      trainedStaffMembers:
        type: string
      updatedAt:
        type: string
      useCases:
        items:
          type: string
        type: array
    type: object
  models.PartnerContact:
    properties:
      businessEmail:
        type: string
      firstName:
        type: string
      jobTitle:
        type: string
      lastName:
        type: string
      phoneNumber:
        type: string
      type:
        description: '"business" or "technical"'
        type: string
    type: object
  models.PartnerRegistrationRequest:
    properties:
      additionalComments:
        type: string
      aiExperience:
        type: string
      annualRevenue:
        type: string
      arangoEngagements:
        type: string
      companyName:
        type: string
      contacts:
        items:
          $ref: '#/definitions/models.PartnerContact'
        minItems: 1
        type: array
      country:
        type: string
      graphDatabaseFit:
        type: string
      industryVerticalFocus:
        items:
          type: string
        minItems: 1
        type: array
      keyPartners:
        type: string
      marketFocus:
        items:
          type: string
        minItems: 1
        type: array
      numberOfEmployees:
        type: string
      primaryPartnerBusiness:
        items:
          type: string
        minItems: 1
        type: array
      registrationStatus:
        type: string
      statusHistoryEntry:
        $ref: '#/definitions/models.StatusHistoryEntry'
      trainedStaffMembers:
        type: string
      useCases:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - aiExperience
    - annualRevenue
    - arangoEngagements
    - companyName
    - contacts
    - country
    - graphDatabaseFit
    - industryVerticalFocus
    - keyPartners
    - marketFocus
    - numberOfEmployees
    - primaryPartnerBusiness
    - trainedStaffMembers
    - useCases
    type: object
  models.StatusHistoryEntry:
    properties:
      changedBy:
        type: string
      comment:
        type: string
      newStatus:
        type: string
      previousStatus:
        type: string
      timestamp:
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: support@partnerportal.com
    name: API Support
  description: API for managing partners, opportunities, products, and accounts in
    the Partner Portal
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Partner Portal API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with username and password, returns JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Invalid request body
          schema:
            type: string
        "401":
          description: Invalid credentials or account inactive
          schema:
            type: string
      summary: User login
      tags:
      - auth
...
  /partners:
    get:
      consumes:
      - application/json
      description: Get paginated list of partners with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 25
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by company name
        in: query
        name: companyName
        type: string
      - description: Filter by country
        in: query
        name: country
        type: string
      - description: Filter by partner type
        in: query
        name: partnerType
        type: string
      - description: Filter by registration status
        in: query
        name: registrationStatus
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated partners
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List all partners
      tags:
      - partners
  /partners/{id}:
    get:
      consumes:
      - application/json
      description: Get a single partner by its ID
      parameters:
      - description: Partner ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Partner'
        "400":
          description: Missing partner ID
          schema:
            type: string
        "404":
          description: Partner not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get partner by ID
      tags:
      - partners
    put:
      consumes:
      - application/json
      description: Update partner details or status
      parameters:
      - description: Partner ID
        in: path
        name: id
        required: true
        type: string
      - description: Partner update data
        in: body
        name: partner
        required: true
        schema:
          $ref: '#/definitions/models.Partner'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Partner'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Partner not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update partner
      tags:
      - partners
  /partners/register:
    post:
      consumes:
      - application/json
      description: Register a new partner (public endpoint, no authentication required)
      parameters:
      - description: Partner registration data
        in: body
        name: partner
        required: true
        schema:
          $ref: '#/definitions/models.PartnerRegistrationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Partner'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Register new partner
      tags:
      - partners
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"